Making __main__ file simple
